ntsc-beta and pal-beta: Decompile mainTick

+++ src/lib/joy.c
@@ -1107,7 +1107,7 @@ s32 joyCountButtonsOnSpecificSamples(u32 *checksamples, s8 contpadnum, u16 mask)
 	return count;
 }
 
-s32 joyGetStickX(s8 contpadnum)
+s8 joyGetStickX(s8 contpadnum)
 {
 	if (g_JoyDataPtr->unk200 < 0 && (g_JoyConnectedControllers >> contpadnum & 1) == 0) {
 		g_JoyBadReadsStickX[contpadnum]++;
@@ -1121,7 +1121,7 @@ s32 joyGetStickX(s8 contpadnum)
 	return g_JoyDataPtr->samples[g_JoyDataPtr->curlast].pads[contpadnum].stick_x;
 }
 
-s32 joyGetStickY(s8 contpadnum)
+s8 joyGetStickY(s8 contpadnum)
 {
 	if (g_JoyDataPtr->unk200 < 0 && (g_JoyConnectedControllers >> contpadnum & 1) == 0) {
 		g_JoyBadReadsStickY[contpadnum]++;
